requirements.txt
pyspark==3.4.1
streamlit==1.28.2
opencv-python==4.8.1.78
numpy==1.26.0
Pillow==10.0.1
